[CHORE] 뉴스 생성 테스트 패키지 위치 변경
